# hotel-backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  hotel-backend_internal_models.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@hotel-system.com
    name: API Support
  description: This is the API documentation for the Hotel Management System.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Hotel Management System API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user with email and password to obtain a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/hotel-backend_internal_models.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful with token and user profile
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid request body or missing credentials
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Invalid email or password
          schema:
            additionalProperties: true
            type: object
      summary: User login
      tags:
      - Auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: Retrieve the profile details of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Profile retrieved
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"